+++ app/src-tauri/src/results/mod.rs
@@ -1,5 +1,0 @@
-pub mod execution_result;
-pub mod execution_result_detail;
-
-pub use execution_result::*;
-pub use execution_result_detail::*;